src/requirements.txt
# requirements.txt
PyVISA
pyusb
PySide6